+++ x/ecocredit/client/basket/tx.go
@@ -19,7 +19,7 @@ import (
 
 const (
 	FlagDisableAutoRetire      = "disable-auto-retire"
-	FlagCreditTypeAbbreviation = "credit-type-abbreviation"
+	FlagCreditTypeAbbrev       = "credit-type-abbrev"
 	FlagAllowedClasses         = "allowed-classes"
 	FlagMinimumStartDate       = "minimum-start-date"
 	FlagStartDateWindow        = "start-date-window"
@@ -36,33 +36,28 @@ func TxCreateBasketCmd() *cobra.Command {
 		Long: strings.TrimSpace(`Creates a bank denom that wraps credits
 
 Parameters:
-		name: the name used to create a bank denom for this basket token.
+
+- name:  the name used to create a bank denom for this basket token
 
 Flags:
-		disable-auto-retire: disables the auto-retirement of credits upon taking credits
-			from the basket. The credits will be auto-retired if disable_auto_retire is
-			false unless the credits were previously put into the basket by the address
-			picking them from the basket, in which case they will remain tradable.
-		credit-type-abbreviation: filters against credits from this credit type abbreviation (e.g. "BIO").
-		allowed_classes: comma separated (no spaces) list of credit classes allowed to be put in
-			the basket (e.g. "C01,C02").
-		min-start-date: the earliest start date for batches of credits allowed into the basket.
-		start-date-window: the duration of time (in seconds) measured into the past which sets a
-			cutoff for batch start dates when adding new credits to the basket.
-		basket-fee: the fee that the curator will pay to create the basket. It must be >= the
-			required Params.basket_creation_fee. We include the fee explicitly here so that the
-			curator explicitly acknowledges paying this fee and is not surprised to learn that the
-			paid a big fee and didn't know beforehand.
-		description: the description to be used in the basket coin's bank denom metadata.`),
-		Example: `
-		$regen tx ecocredit create-basket HEAED
-			--from regen...
-			--credit-type-abbreviation=FOO
-			--allowed_classes="class1,class2"
-			--basket-fee=100regen
-			--description="any description"
-		`,
-		Args: cobra.ExactArgs(1),
+
+- disable-auto-retire: disables the auto-retirement of credits upon taking credits
+    from the basket. The credits will be auto-retired if disable_auto_retire is
+    false unless the credits were previously put into the basket by the address
+    picking them from the basket, in which case they will remain tradable.
+- credit-type-abbrev: filters against credits from this credit type abbreviation (e.g. "BIO").
+- allowed_classes: comma separated (no spaces) list of credit classes allowed to be put in
+    the basket (e.g. "C01,C02").
+- min-start-date: the earliest start date for batches of credits allowed into the basket.
+- start-date-window: the duration of time (in seconds) measured into the past which sets a
+    cutoff for batch start dates when adding new credits to the basket.
+- basket-fee: the fee that the curator will pay to create the basket. It must be >= the
+    required Params.basket_creation_fee. We include the fee explicitly here so that the
+    curator explicitly acknowledges paying this fee and is not surprised to learn that they
+    paid a big fee and didn't know beforehand.
+- description: the description to be used in the basket coin's bank denom metadata.`),
+		Example: `regen tx ecocredit create-basket NCT --credit-type-abbrev C --allowed_classes C01,C02 basket-fee 100000000uregen description "NCT basket"`,
+		Args:    cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx, err := client.GetClientTxContext(cmd)
 			if err != nil {
@@ -74,7 +69,7 @@ Flags:
 				return err
 			}
 
-			creditTypeName, err := cmd.Flags().GetString(FlagCreditTypeAbbreviation)
+			creditTypeName, err := cmd.Flags().GetString(FlagCreditTypeAbbrev)
 			if err != nil {
 				return err
 			}
@@ -158,7 +153,7 @@ Flags:
 
 	// command flags
 	cmd.Flags().Bool(FlagDisableAutoRetire, false, "dictates whether credits will be auto-retired upon taking")
-	cmd.Flags().String(FlagCreditTypeAbbreviation, "", "filters against credits from this credit type abbreviation (e.g. \"C\")")
+	cmd.Flags().String(FlagCreditTypeAbbrev, "", "filters against credits from this credit type abbreviation (e.g. \"C\")")
 	cmd.Flags().StringSlice(FlagAllowedClasses, []string{}, "comma separated (no spaces) list of credit classes allowed to be put in the basket (e.g. \"C01,C02\")")
 	cmd.Flags().String(FlagMinimumStartDate, "", "the earliest start date for batches of credits allowed into the basket (e.g. \"2012-01-01\")")
 	cmd.Flags().Uint64(FlagStartDateWindow, 0, "sets a cutoff for batch start dates when adding new credits to the basket (e.g. 1325404800)")
@@ -166,7 +161,7 @@ Flags:
 	cmd.Flags().String(FlagDenomDescription, "", "the description to be used in the bank denom metadata.")
 
 	// required flags
-	cmd.MarkFlagRequired(FlagCreditTypeAbbreviation)
+	cmd.MarkFlagRequired(FlagCreditTypeAbbrev)
 	cmd.MarkFlagRequired(FlagAllowedClasses)
 
 	return txFlags(cmd)
@@ -174,19 +169,17 @@ Flags:
 
 func TxPutInBasketCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "put-in-basket [basket_denom] [credits]",
-		Short: "add credits to the basket",
-		Long: strings.TrimSpace(`add credits to the basket.
-Parameters:
-		basket_denom: basket identifier
-		credits: path to JSON file containing credits to put in the basket
-Flags:
-		from: account address of the owner
-		`),
-		Example: `
-regen tx ecocredit put-in-basket eco.uC.NCT credits.json
+		Use:   "put-in-basket [basket-denom] [credits-json]",
+		Short: "Add credits to a basket",
+		Long: strings.TrimSpace(`Add credits to a basket.
 
-Where the credits.json file contains:
+Parameters:
+
+- basket-denom:  basket identifier
+- credits-json:  path to JSON file containing credits to put in the basket`),
+		Example: `regen tx ecocredit put-in-basket eco.uC.NCT credits.json
+
+Example JSON:
 
 [
 	{
@@ -197,8 +190,7 @@ Where the credits.json file contains:
 		"batch_denom": "C01-001-20210101-20220101-001",
 		"amount": "10.5"
 	}
-]
-		`,
+]`,
 		Args: cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx, err := client.GetClientTxContext(cmd)
@@ -230,25 +222,25 @@ Where the credits.json file contains:
 
 func TxTakeFromBasketCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "take-from-basket [basket_denom] [amount]",
+		Use:   "take-from-basket [basket-denom] [amount]",
 		Short: "Takes credits from a basket",
-		Long: strings.TrimSpace(`takes credits from a basket starting from the oldest credits first.
+		Long: strings.TrimSpace(`Takes credits from a basket starting from the oldest credits first.
+
 Parameters:
-		basket_denom: denom identifying basket from which we redeem credits.
-		amount: amount is a positive integer number of basket tokens to convert into credits.
+
+- basket-denom:  denom identifying basket from which we redeem credits.
+- amount:        number of basket tokens to convert into credits.
+
 Flags:
-		from: account address of the owner of the basket.
-		retirement-jurisdiction: retirement jurisdiction is the optional retirement jurisdiction for the credits
-				which will be used only if --retire-on-take flag is true.
-		retire-on-take: retire on take is a boolean that dictates whether the ecocredits
-		                received in exchange for the basket tokens will be received as
-		                retired or tradable credits.
+
+- retirement-jurisdiction: optional retirement jurisdiction
+    for the credits which will be used only if --retire-on-take flag is true.
+- retire-on-take: boolean that dictates whether the ecocredits received
+    in exchange for the basket tokens will be received as retired or tradable credits.
 		
 		`),
-		Example: `
-regen tx ecocredit take-from-basket eco.uC.NCT 1000
-regen tx ecocredit take-from-basket eco.uC.NCT 1000 --retire-on-take true --retirement-jurisdiction AQ
-		`,
+		Example: `regen tx ecocredit take-from-basket eco.uC.NCT 1000
+regen tx ecocredit take-from-basket eco.uC.NCT 1000 --retire-on-take=true --retirement-jurisdiction "US-WA 98225"`,
 		Args: cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx, err := client.GetClientTxContext(cmd)