+++ CHANGELOG.md
@@ -27,6 +27,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#1274](https://github.com/regen-network/regen-ledger/pull/1274) Add `cancel-sell-order` command
 
+#### Changed
+
+- [#1288](https://github.com/regen-network/regen-ledger/pull/1288) Rename `types` query command to `credit-types`
+- [#1288](https://github.com/regen-network/regen-ledger/pull/1288) Rename `balance` query command to `batch-balance`
+- [#1288](https://github.com/regen-network/regen-ledger/pull/1288) Rename `supply` query command to `batch-supply`
+- [#1288](https://github.com/regen-network/regen-ledger/pull/1288) Rename `buy-direct-batch` transaction command to `buy-direct-bulk`
+
 #### Fixed
 
 - [#1278](https://github.com/regen-network/regen-ledger/pull/1278) Fix `open` not set in `Msg/CreateBatch`